[Fix] Fix size_t to UINT narrowing warnings and add KojeomEditor to solution config
- Add static_cast<UINT>() in Texture.cpp CopyPixels and SysMemPitch to eliminate C4267 warnings
- Add KojeomEditor project configuration platform mappings to solution file so it builds with the solution

+++ Engine/Graphics/Texture.cpp
@@ -91,7 +91,7 @@ HRESULT KTexture::LoadFromFile(ID3D11Device* Device, const std::wstring& Filenam
 
     std::vector<BYTE> pixelData(static_cast<size_t>(imageSize));
 
-    hr = converter->CopyPixels(nullptr, stride, imageSize, pixelData.data());
+    hr = converter->CopyPixels(nullptr, static_cast<UINT>(stride), static_cast<UINT>(imageSize), pixelData.data());
     if (FAILED(hr))
     {
         KLogger::HResultError(hr, "Failed to copy WIC pixel data");
@@ -116,7 +116,7 @@ HRESULT KTexture::LoadFromFile(ID3D11Device* Device, const std::wstring& Filenam
 
     D3D11_SUBRESOURCE_DATA initData = {};
     initData.pSysMem = pixelData.data();
-    initData.SysMemPitch = stride;
+    initData.SysMemPitch = static_cast<UINT>(stride);
     initData.SysMemSlicePitch = 0;
 
     hr = Device->CreateTexture2D(&textureDesc, &initData, &Texture);
